+++ src/structures/battlesnake.rs
@@ -6,7 +6,7 @@ use crate::coordinate::Coordinate;
 // Define the Battlesnake struct
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Battlesnake {
-    id: i32,
+    id: u8,
     health: i32,
     body: VecDeque<Coordinate>,
     latency: i32,
@@ -17,13 +17,13 @@ pub struct Battlesnake {
 impl Battlesnake {
     // Battlesnake constructor
     pub fn new(
-        id: i32, health: i32, body: VecDeque<Coordinate>, latency: i32,
+        id: u8, health: i32, body: VecDeque<Coordinate>, latency: i32,
         head: Coordinate, length: usize
     ) -> Battlesnake {
         Battlesnake {id, health, body, latency, head, length}
     }
 
-    pub fn get_id(&self) -> i32 {
+    pub fn get_id(&self) -> u8 {
         self.id
     }
